Refactoring connection -> edge
