+++ src/rust/lib/graph-editor/src/component.rs
@@ -1,7 +1,11 @@
 //! Root module for graph component definitions.
 
-pub mod connection;
-pub mod connection2;
+pub mod edge;
+pub mod edge2;
 pub mod cursor;
 pub mod node;
 pub mod visualization;
+
+pub use cursor::Cursor;
+pub use edge::Edge;
+pub use node::Node;
